feat: pg-simulator 모듈 추가 및 주문 요청 시 결제 요청 기능 추가

+++ apps/commerce-api/src/main/java/com/loopers/application/product/ProductResult.java
@@ -10,7 +10,7 @@ public record ProductResult(Long id,
                             Long price,
                             String mainImageUrl,
                             String description,
-                            Integer stockQuantity,
+                            Long stockQuantity,
                             Integer likeCount) implements Serializable {
     public static ProductResult from(Product product) {
         return new ProductResult(
